requirements.txt

openai
upbit-client
python-telegram-bot
streamlit
pandas
numpy
matplotlib
scikit-learn
python-dotenv

